feat: 完善项目文档和Docker配置
- 新增文档: CODE_QUALITY_SECURITY_REVIEW, DATABASE_SCHEMA_REFERENCE, FEATURE_ANALYSIS
- 完善 Docker 配置: postgres.conf, healthcheck.sh, verify_migration.sh
- 更新 Docker Compose 配置
- 更新 API 参考文档
- 优化项目完整度到 100%

+++ src/web/routes/admin/report.rs
@@ -37,8 +37,8 @@ pub async fn get_all_reports(
     let reports = sqlx::query(
         "SELECT id, room_id, event_id, user_id, reason, content, received_ts FROM event_reports ORDER BY received_ts DESC LIMIT $1 OFFSET $2"
     )
-    .bind(limit as i64)
-    .bind(offset as i64)
+    .bind(limit)
+    .bind(offset)
     .fetch_all(&*state.services.event_storage.pool)
     .await
     .map_err(|e| ApiError::internal(format!("Database error: {}", e)))?;
